design #262 : 홀로그램 수정
