+++ Outline/Outline/Source/Model/CardModel.swift
@@ -1,0 +1,42 @@
+//
+//  CardModel.swift
+//  Outline
+//
+//  Created by hyunjun on 11/19/23.
+//
+
+import SwiftUI
+
+enum CardType {
+    case freeRun
+    case nice
+    case good
+    case excellent
+    
+    var cardFrondSideImage: String {
+        switch self {
+        case .nice, .freeRun:
+            "NormalCardFrontSide"
+        case .good, .excellent:
+            "CardFrontSide"
+        }
+    }
+    
+    var cardBackSideImage: String {
+        switch self {
+        case .freeRun, .nice:
+            "NormalCardBackSide"
+        case .good, .excellent:
+            "CardBackSide"
+        }
+    }
+    
+    var hologramImage: String {
+        switch self {
+        case .freeRun, .nice, .good:
+            "Hologram1"
+        case .excellent:
+            "Hologram2"
+        }
+    }
+}
